third_party/blink/web_tests/http/tests/xmlhttprequest/xmlhttprequest-inheritance-expected.txt

Checks the prototype chain of XMLHttpRequest / XMLHttpRequestUpload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS XMLHttpRequest.__proto__.name is "XMLHttpRequestEventTarget"
PASS XMLHttpRequest.prototype.__proto__.constructor.name is "XMLHttpRequestEventTarget"
FAIL className(XMLHttpRequest.prototype.__proto__) should be XMLHttpRequestEventTargetPrototype. Was XMLHttpRequestEventTarget.
PASS XMLHttpRequest.prototype.__proto__.__proto__ is EventTarget.prototype
PASS XMLHttpRequestUpload.__proto__.name is "XMLHttpRequestEventTarget"
PASS XMLHttpRequestUpload.prototype.__proto__.constructor.name is "XMLHttpRequestEventTarget"
FAIL className(XMLHttpRequestUpload.prototype.__proto__) should be XMLHttpRequestEventTargetPrototype. Was XMLHttpRequestEventTarget.
PASS XMLHttpRequestUpload.prototype.__proto__.__proto__ is EventTarget.prototype
PASS successfullyParsed is true

TEST COMPLETE
